new 20min commit
i had an error, when the battle logic played out, the opp health had the same health as me. it took me a bit, but i realized i was printing out the userhealth in the opp portion. i also created a scrollable box to look at the move list. impleneted the addmovemessage, which is just appending the message. lastly, the gameresult logic. it was pretty simple- just checks whose health is less than or equal to 0