fixing healthbar bugs + transition to results
added logic, checks if gameResult is not nil. if not nil, then its time to navigate to next view.

noticed that when running the code, the bar isnt actually that accurate. realized i was dividing CGFloat by a factor of 100. it should be 200 because the health of each pkmn is set to 200. also for the healthcolor. for percentage, its to be divded by 200, not 100. i thougt it was supposed to be 100 for both values bc 100 always seems like the default value.